+++ src/models/message_response.rs
@@ -10,7 +10,6 @@ pub struct MessageResponse {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct MessageStatusResponse {
     pub success: Option<bool>,
-    // TODO: error and otehr attributes
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]